tons of played button adds
