+++ Jellyfin.UWP/Controls/ViewedFavoriteButtonControl.xaml
@@ -19,11 +19,11 @@
 
     <Canvas Background="Transparent">
         <StackPanel Canvas.Left="{x:Bind PositionLeft}" Canvas.Top="{x:Bind PositionTop}" Canvas.ZIndex="2" Orientation="Horizontal">
-            <Button x:Name="btn_Viewed" Margin="0,0,5,0" Command="{Binding PlayedStateCommand}">
+            <Button x:Name="btn_Viewed" Margin="0,0,5,0" Click="btn_Viewed_Click">
                 <Button.Background>
                     <SolidColorBrush Color="#000000" Opacity="0" />
                 </Button.Background>
-                <FontIcon Foreground="{Binding HasBeenViewed, Converter={StaticResource TrueRedFalseWhite}}" FontFamily="Segoe MDL2 Assets" Glyph="&#xE73E;" />
+                <FontIcon Foreground="{x:Bind Item.UserData.HasBeenWatched, Converter={StaticResource TrueRedFalseWhite}}" FontFamily="Segoe MDL2 Assets" Glyph="&#xE73E;" />
             </Button>
 
             <Button x:Name="btn_Favorite" Margin="5,0,0,0" Click="btn_Favorite_Click">
