+++ Jellyfin.UWP/Controls/ViewedIconControl.xaml
@@ -1,0 +1,22 @@
+﻿<UserControl
+    x:Class="Jellyfin.UWP.Controls.ViewedIconControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:converters="using:Microsoft.Toolkit.Uwp.UI.Converters"
+    mc:Ignorable="d"
+    d:DesignHeight="300"
+    d:DesignWidth="400">
+    
+    <UserControl.Resources>
+        <converters:BoolToVisibilityConverter x:Key="BoolToVisibilityConverter" />
+    </UserControl.Resources>
+
+    <Canvas Background="Transparent">
+        <Grid Canvas.Left="{x:Bind PositionLeft}" Canvas.Top="{x:Bind PositionTop}" Canvas.ZIndex="2" Visibility="{x:Bind HasBeenViewed, Converter={StaticResource BoolToVisibilityConverter}}">
+            <FontIcon Foreground="#00a4dc" FontFamily="Segoe MDL2 Assets" Glyph="&#xF63C;" />
+            <FontIcon FontFamily="Segoe MDL2 Assets" Glyph="&#xE73E;" />
+        </Grid>
+    </Canvas>
+</UserControl>
